Missed imports Fix open_file
