+++ pyproject.toml
@@ -102,6 +102,7 @@ doc = [
     "sphinx-gallery>=0.14.0,<0.20",
     "sphinx_design>=0.4.0,<0.7",
     "pyvista[io]>=0.38.0,<0.45",
+    "ansys-tools-visualization-interface; python_version >= '3.10'",
 ]
 all = [
     "matplotlib>=3.5.0,<3.11",
